Regression Model on Market Segmentation
Course Advanced Statistics
This project identified the factors to study the satisfaction level of customers. A regression model to be built using these factors for Market segmentation. The objective of the project is to use the dataset to build a regression model to predict customer satisfaction.
Skills and Tools
Correlation, Regression, Multicollinearity, PCA/FA